feat: separate Modal component from ModalWithForm component
